+++ tests/test-runners/jest/test.js
@@ -1,7 +1,0 @@
-describe('Realm', function() {
-   it('should be requirable', function() {
-       var realm = require('realm');
-       expect(typeof realm).toBe('function');
-       expect(realm.name).toBe('Realm');
-   });
-});